… src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarHelpers.Unix.cs
… src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarHelpers.cs
… src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarReader.cs
… src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarWriter.Unix.cs
… src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarWriter.Windows.cs
… src/libraries/System.Formats.Tar/src/System/Formats/Tar/TarWriter.cs
… src/libraries/System.IO.Compression.ZipFile/src/System/IO/Compression/ZipFile.Create.cs