Low-hanging fruit.

|    Method |         Toolchain | format |      Mean | Ratio | Code Size | Allocated | Alloc Ratio |
|---------- |------------------ |------- |----------:|------:|----------:|----------:|------------:|
| Roundtrip | \main\corerun.exe |  Ustar | 12.630 us |  1.00 |  11,006 B |   17.1 KB |        1.00 |
| Roundtrip |   \pr\corerun.exe |  Ustar |  8.266 us |  0.65 |   7,526 B |   5.95 KB |        0.35 |
|           |                   |        |           |       |           |           |             |
| Roundtrip | \main\corerun.exe |    Pax | 51.859 us |  1.00 |  15,875 B |  83.99 KB |        1.00 |
| Roundtrip |   \pr\corerun.exe |    Pax | 31.182 us |  0.60 |  11,707 B |  31.62 KB |        0.38 |
|           |                   |        |           |       |           |           |             |
| Roundtrip | \main\corerun.exe |    Gnu | 14.264 us |  1.00 |  13,310 B |  19.76 KB |        1.00 |
| Roundtrip |   \pr\corerun.exe |    Gnu |  8.183 us |  0.57 |   8,389 B |   5.95 KB |        0.30 |

```C#
private MemoryStream _archive = new MemoryStream();
private string[] _names;
private MemoryStream[] _streams;

[GlobalSetup]
public void Setup()
{
    _names = Enumerable.Range(0, 10).Select(i => $"HelloWorld{i}.txt").ToArray();
    _streams = _names.Select(s => new MemoryStream(Encoding.UTF8.GetBytes(s))).ToArray();
}

[Benchmark]
[Arguments(TarEntryFormat.Pax)]
[Arguments(TarEntryFormat.Gnu)]
[Arguments(TarEntryFormat.Ustar)]
public void Roundtrip(TarEntryFormat format)
{
    using (TarWriter writer = new TarWriter(_archive, leaveOpen: true))
    {
        for (int i = 0; i < _names.Length; i++)
        {
            _streams[i].Position = 0;
            TarEntry entry = format switch
            {
                TarEntryFormat.Pax => new PaxTarEntry(TarEntryType.RegularFile, _names[i]) { DataStream = _streams[i] },
                TarEntryFormat.Gnu => new GnuTarEntry(TarEntryType.RegularFile, _names[i]) { DataStream = _streams[i] },
                _ => new UstarTarEntry(TarEntryType.RegularFile, _names[i]) { DataStream = _streams[i] },
            };
            writer.WriteEntry(entry);
        }
    }

    _archive.Position = 0;

    using (TarReader reader = new TarReader(_archive, leaveOpen: true))
    {
        TarEntry entry;
        while ((entry = reader.GetNextEntry()) != null)
        {
            entry.DataStream?.CopyTo(Stream.Null);
        }
    }
}
```