@carlossanlop 2022-08-21
/backport to release/7.0-rc1